move function into utils
